requirements.txt
pandas==2.2.1
psycopg2-binary==2.9.9
PyYAML==6.0.1
SQLAlchemy==2.0.28

